ovisbot/requirements.txt
aiohttp==3.5.4
argh==0.26.2
async-timeout==3.0.1
attrs==19.1.0
BTrees==4.5.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorthief==0.2.1
cycler==0.10.0
discord==1.0.1
discord.py==1.1.1
idna==2.8
idna-ssl==1.1.0
kiwisolver==1.1.0
multidict==4.5.2
numpy==1.16.4
pathtools==0.1.2
persistent==4.5.0
Pillow==6.2.0
pycparser==2.19
pycrypto==2.6.1
pymodm==0.4.1
pymongo==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1
requests==2.22.0
six==1.12.0
transaction==2.4.0
typing-extensions==3.7.2
urllib3==1.25.3
websockets==6.0
yarl==1.3.0
zc.lockfile==1.4
ZConfig==3.4.0
zodbpickle==1.0.3
zope.interface==4.6.0
feedparser==5.2.1
